+++ views/pages/infoNasa.ejs
@@ -1,6 +1,6 @@
 <head>
   <% include ../partials/head %>
-  <link href="../../public/projects.css" rel="stylesheet" type="text/css" />
+  <link href="/projects.css" rel="stylesheet" type="text/css" />
 </head>
 <body>
   <div class="container">
@@ -9,12 +9,7 @@
       <div class="col-md-12">
         <div class="row">
           <div class="col-md-4">
-            <img
-              class="space-img"
-              src="<%= data[i].image_url %>"
-              alt=""
-              style="width: 100%; height: auto%; object-fit:cover"
-            />
+            <img class="space-img" src="<%= data[i].image_url %>" alt="" />
           </div>
           <div class="col-md-8">
             <h3 class="space-title"><%= data[i].name %></h3>
@@ -26,6 +21,7 @@
         </div>
       </div>
     </div>
+    <div class="separator"></div>
     <% } %>
   </div>
 </body>
