+++ views/partials/head.ejs
@@ -7,6 +7,7 @@
   rel="stylesheet"
   href="//maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css"
 />
+<link href="../../public/projects.css" rel="stylesheet" type="text/css" />
 <style>
   body {
     padding-top: 50px;
